refactor: move RecipeDetails into pages and refactor RecipeCard component
